requirements.txt
opencv-python>=4.9.0
pytesseract>=0.3.10
imutils>=0.5.4
pdf2image>=1.16.3
Pillow>=10.0.0
pandas>=2.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
PyYAML>=6.0
click>=8.1.0
numpy>=1.26.0
streamlit>=1.30.0